trial data
There are 7 DFAs.  The .jff files are JFLAP.  The language.py file
contains definitions of the DFAs and code that will generate strings
that are within the language of the DFAs.  The language.py file is
currently configured to generate 1000 samples of each DFA and then test
to ensure that there is no cross acceptance of the strings with any
other DFA.  That is, the strings generated for DFA_0 will not be
accepted by any other DFA.  If there is an acceptance the message,
"error" is printed.  It is up to the user to rerun language.py until no
"error" strings are printed.

trial_data.csv contains a successful run of the language.py file with
the 7 DFAs.