Updated so the RNN will run.
There is something wrong with the RNN itself.  It doesn't seem to be computing its accuracy correctly.  Need to look in the line 124-136 area.

The data files have changed to string lengths of 32 for the time being.  I cannot re-upload those because of a file size limiitation on GitHub.  I'll share these via google drive.